+++ HISTORY.md
@@ -1,6 +1,10 @@
 ## [0.3.0][] • work in progress
 
 - Updated to Swift 2.1 (Xcode 7.1).
+- Added default `DOTLabel` implementation (`return "\(self)"`).  You don't
+  have to change anything.  If you have custom `DOTLabel` implementations,
+  you can keep using them, but if the only thing you do is `case Foo: return
+  "Foo"`, you can safely delete this code.
 
 
 ## [0.2.0][] • 2015-06-06